ref(typescript): remove type
