+++ static/app/components/profiling/flamegraphTooltip.tsx
@@ -5,11 +5,11 @@ import {t} from 'sentry/locale';
 import {Flamegraph} from 'sentry/utils/profiling/flamegraph';
 import {FlamegraphFrame} from 'sentry/utils/profiling/flamegraphFrame';
 import {TypeScriptProfile} from 'sentry/utils/profiling/profile/formats/chromeTraceProfile';
-import {getResolvedTypescriptTypeName} from 'sentry/utils/profiling/profile/typescript/importTypeScriptTypesJSON';
+import {getResolvedTypescriptTypeName} from 'sentry/utils/profiling/profile/formats/typescript/importTypeScriptJSON';
 
 interface TypeScriptTooltipProps {
   frame: FlamegraphFrame;
-  tree: TypeScriptTypes.TypeTree | undefined;
+  tree: TypeScript.TypeTree | undefined;
   flamegraph?: Flamegraph;
 }
