+++ static/app/utils/profiling/profile/formats/typescript/samples/program.ts
