+++ static/app/utils/profiling/profile/formats/typescript/samples/tsc_trace/trace.json
