+++ static/app/utils/profiling/profile/formats/typescript/samples/tsc_trace/types.json
