+++ tests/js/spec/utils/profiling/profile/guards.spec.tsx
